Moving handle classes into common project
