+++ src/Microsoft.Data.SqlClient/src/Interop/Windows/Handles/DebugCriticalHandleMinusOneIsInvalid.netcore.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#if DEBUG && !NET8_0_OR_GREATER
+#if DEBUG && !NETFRAMEWORK && !NET8_0_OR_GREATER
 
 using Microsoft.Win32.SafeHandles;
