+++ src/Microsoft.Data.SqlClient/src/Interop/Windows/Handles/DebugCriticalHandleZeroOrMinusOneIsInvalid.netcore.cs
@@ -2,13 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#if DEBUG && !NET8_0_OR_GREATER
+#if DEBUG && !NETFRAMEWORK && !NET8_0_OR_GREATER
 
-using System.Net.NetworkInformation;
-using System.Net.Sockets;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Threading;
 using Microsoft.Win32.SafeHandles;
 
 namespace System.Net
@@ -17,36 +12,33 @@ namespace System.Net
     // This is a helper class for debugging GC-ed handles that we define.
     // As a general rule normal code path should always destroy handles explicitly
     //
-    internal abstract class DebugSafeHandle : SafeHandleZeroOrMinusOneIsInvalid
+    internal abstract class DebugCriticalHandleZeroOrMinusOneIsInvalid : CriticalHandleZeroOrMinusOneIsInvalid
     {
         private string _trace;
 
-        protected DebugSafeHandle(bool ownsHandle) : base(ownsHandle)
+        protected DebugCriticalHandleZeroOrMinusOneIsInvalid() : base()
         {
-            Trace();
-        }
-
-        protected DebugSafeHandle(IntPtr invalidValue, bool ownsHandle) : base(ownsHandle)
-        {
-            SetHandle(invalidValue);
             Trace();
         }
 
         private void Trace()
         {
-            _trace = "WARNING! GC-ed  >>" + this.GetType().ToString() + "<< (should be explicitly closed) \r\n";
+            _trace = "WARNING! GC-ed  >>" + this.GetType().FullName + "<< (should be explicitly closed) \r\n";
+            if (NetEventSource.IsEnabled)
+                NetEventSource.Info(this, "Creating SafeHandle");
 #if TRACE_VERBOSE
             string stacktrace = Environment.StackTrace;
             _trace += stacktrace;
-#endif
+#endif //TRACE_VERBOSE
         }
 
-        ~DebugSafeHandle()
+        ~DebugCriticalHandleZeroOrMinusOneIsInvalid()
         {
+            DebugThreadTracking.SetThreadSource(ThreadKinds.Finalization);
             if (NetEventSource.IsEnabled)
                 NetEventSource.Info(this, _trace);
         }
     }
 }
 
-#endif
+#endif // !NET8_0_OR_GREATER
